security-agent check can now send an event
